+++ NEWS.md
@@ -1,4 +1,4 @@
-# chevron 0.2.0.9014
+# chevron 0.2.0.9017
 
 * Add `row_split_var` and `page_var` as argument for some template.
 * Update `LBT06` template to use `PARAM` as row split.
@@ -8,7 +8,8 @@
 * Add `RMPT03` template.
 * Update `EGT03` to use `ACTARMCD` as default arm variable, and remove the preprocessing of filtering to "HR".
 * Update `EXT01` to allow it to be displayed by `PARCAT2`.
-* Add more templates: `AET05` and `AET05_ALL`.
+* Add `COXT01` template. `COXT02` is based on `COXT01` now.
+* Add `AET05` and `AET05_ALL` template.
 
 # chevron 0.2.0